+++ packages/executor-react/src/hooks/useAddTxIntention.ts
@@ -54,8 +54,20 @@ export const useAddTxIntention = ({
 	const { intentions = [] } = useStore(customStore);
 
 	const { mutate, mutateAsync, ...rest } = useMutation({
-		mutationFn: ({ reset, publicKey, intention }: AddTxIntentionVariables) => {
-			return addTxIntention(config, store, intention, reset, publicKey);
+		mutationFn: async ({
+			reset,
+			publicKey,
+			intention,
+		}: AddTxIntentionVariables) => {
+			const int = await addTxIntention(config, intention, publicKey);
+
+			store.setState((state) => {
+				return {
+					intentions: reset ? [int] : [...(state.intentions || []), int],
+				};
+			});
+
+			return int;
 		},
 	});