+++ packages/executor/src/actions/index.ts
@@ -1,5 +1,4 @@
 export * from "./addTxIntention";
-export * from "./clearTxIntentions";
 export * from "./finalizeBTCTransaction";
 export * from "./getBTCFeeRate";
 export * from "./getPublicKey";